NYC_Scraper/nyc_apartments/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc-apartments"
version = "0.1.0"
description = "NYC apartment-like listings scraper and aggregator"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
keywords = ["nyc", "apartments", "scraper", "cli"]

dependencies = [
    "requests",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "typer[all]",
    "rich",
    "geopy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "mypy",
    "ruff",
]

[project.scripts]
nyc-apartments = "nyc_apartments.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
